rework how we treat CJK (full width) characters
previously, full-width CJK characters were only handled in the renderer, which led to some workarounds and made syntax highlighting tricky. This changes it so we account for full-width characters in our data model directly (e.g. Screen/Row/Cell) which leads to a simpler renderer and solves for syntax highlighting and word wrapping edge cases.