Add workaround for garbled utf-8 characters on linux
Fixes https://github.com/waf/CSharpRepl/issues/261
See original fix in https://github.com/waf/PrettyPrompt/pull/231 that
introduced this problem.

Ultimately it's a workaround for a dotnet bug:
https://github.com/dotnet/runtime/issues/88343

We can run the "Windows Only" logic on Linux once the above bug is
fixed, but there's no ETA for it.